+++ src/app/core/authentication/authentication.service.ts
@@ -322,6 +322,10 @@ export class AuthenticationService {
     this.setCredentials(credentials);
   }
 
+  public saveZitadeloAuthTokenDetailsStorageKey(tokenResponse: OAuth2Token): void {
+    this.storage.setItem(this.oAuthTokenDetailsStorageKey, JSON.stringify(tokenResponse));
+  }
+
   /**
    * Following functions are for two factor authentication and require
    * first level authorization headers to be setup for the requests.
