+++ src/app/notifications/notifications.service.ts
@@ -4,6 +4,7 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 
 /** rxjs Imports */
 import { Observable, of } from 'rxjs';
+import { AuthService } from 'app/zitadel/auth.service';
 
 /**
  * Notification service.
@@ -15,12 +16,19 @@ export class NotificationsService {
   /**
    * @param {HttpClient} http Http Client to send requests.
    */
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    private authService: AuthService
+  ) {}
 
   /**
    * @returns {Observable<any>} Notifications.
    */
   getNotifications(isRead: boolean, limit: number): Observable<any> {
+    if (localStorage.getItem('mifosXZitadel') === 'true') {
+      this.authService.notification();
+      return;
+    }
     const httpParams = new HttpParams().set('isRead', isRead.toString()).set('limit', limit);
     return this.http.get('/notifications', { params: httpParams });
   }
@@ -29,6 +37,10 @@ export class NotificationsService {
    * @returns {Observable<any>} Notifications.
    */
   updateNotifications(): Observable<any> {
+    if (localStorage.getItem('mifosXZitadel') === 'true') {
+      this.authService.notification();
+      return;
+    }
     return this.http.put('/notifications', {});
   }
