+++ src/app/zitadel/users/view-user/view-user.component.scss
@@ -1,0 +1,14 @@
+.container {
+  max-width: 37rem;
+
+  .content {
+    div {
+      margin: 1rem 0;
+      word-wrap: break-word;
+
+      span {
+        display: block;
+      }
+    }
+  }
+}
