+++ src/app/zitadel/users/view-user/view-user.component.spec.ts
@@ -1,0 +1,24 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ViewUserZitadelComponent } from './view-user.component';
+
+describe('ViewUserComponent', () => {
+  let component: ViewUserZitadelComponent;
+  let fixture: ComponentFixture<ViewUserZitadelComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ViewUserZitadelComponent]
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ViewUserZitadelComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
